Refactored Controller Logic
* Refactored ToPDF() function so Submission.cs handles the majority of
  work and delegates the individual entry tables to its children
  Timesheet.cs and MileageForm.cs
* Combined ModalProcess() and Process() into one function that returns
  different views depending on where it was called from
