+++ AdminUI/Views/Home/_TimesheetTable.cshtml
@@ -323,9 +323,10 @@
                         </div>
                     </div>
                     <div class="modal-footer @("modal-footer" + @t.Id)">
-                        <form asp-controller="Submission" asp-action="ModalProcess">
+                        <form asp-controller="Submission" asp-action="Process">
                             <a class="btn btn-secondary" asp-controller="Submission" asp-action="Index" asp-route-id="@t.Id" title="Opens a full-detailed page on this submission">More Details</a>
                             <input type="hidden" name="id" value="@t.Id"/>
+                            <input type="hidden" name="modal" value="true"/>
                             <input class="btn btn-primary" name="Status" value="Approve" type="submit" formnovalidate>
                             <input class="btn btn-danger" name="Status" value="Reject" type="submit">
                             <input type="text" name="RejectionReason" placeholder="Rejection Reason" required/>
